Update300924

+++ Semeantica/prueba.cpp
@@ -2,7 +2,7 @@ static void Main(string[] args)
 {
     char a,b,c,d;
     float pi,e;
-    a=(int)((3+5)*8-(10-4)/2);
+    a=((3+5)*8-(10-4)/2);
     d = (34 + 1);
     a = 255;
     b++;
